give permission to product controller for part of => index/store/update/delete/restore/uploade image

+++ app/Http/Controllers/ProductController.php
@@ -33,46 +33,62 @@ class ProductController extends Controller
 
     public function uploadImage(Request $request, $id)
     {
-        $request->validate([
-            'image' => 'required|max:10000|file|mimes:jpg,png,jpeg'
-        ]);
-        $product = Product::findOrFail($id);
-        if ($request->hasFile('image')) {
-            if ($product->getFirstMedia('images')) {
-                $product->clearMediaCollection('images');
-            }
-            $media = $product->addMedia($request->file('image'))->toMediaCollection('images');
-            $mediaUrl = $media->getUrl(); // URL to access the image
-            $mediaName = $media->name; // Optional: Media name (if set)
-
-            return response()->json([
-                'message' => 'تصویر با موفقیت اپلود شد',
-                'media_url' => $mediaUrl,
-                'media_name' => $mediaName,
+        if ($request->user()->can('product.uploadimage')) {
+            $request->validate([
+                'image' => 'required|max:10000|file|mimes:jpg,png,jpeg'
             ]);
+            $product = Product::findOrFail($id);
+            if ($request->hasFile('image')) {
+                if ($product->getFirstMedia('images')) {
+                    $product->clearMediaCollection('images');
+                }
+                $media = $product->addMedia($request->file('image'))->toMediaCollection('images');
+                $mediaUrl = $media->getUrl(); // URL to access the image
+                $mediaName = $media->name; // Optional: Media name (if set)
+
+                return response()->json([
+                    'message' => 'تصویر با موفقیت اپلود شد',
+                    'media_url' => $mediaUrl,
+                    'media_name' => $mediaName,
+                ]);
+            } else {
+                return response()->json(['message' => 'هیچ تصویری برای اپلود یافت نشد'], 400);
+            }
         } else {
-            return response()->json(['message' => 'هیچ تصویری برای اپلود یافت نشد'], 400);
+            return response()->json(['message' => 'شما دسترسی انجام این کار را ندارید'], 403);
         }
     }
 
     public function update(UpdateProductRequest $request, $id)
     {
-        $product = Product::findOrFail($id);
-        $product->update($request->validated());
-        return response()->json(['message' => 'محصول با موفقیت به روز رسانی شد', 'product' => $product]);
+        if ($request->user()->can('product.update')) {
+            $product = Product::findOrFail($id);
+            $product->update($request->validated());
+            return response()->json(['message' => 'محصول با موفقیت به روز رسانی شد', 'product' => $product]);
+        } else {
+            return response()->json(['message' => 'شما دسترسی لازم برای انجام این کار را ندارید'], 403);
+        }
     }
 
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        $product = Product::findOrFail($id);
-        $product->delete();
-        return response()->json(['محصول با موفقیت حذف شد ']);
+        if ($request->user()->can('product.delete')) {
+            $product = Product::findOrFail($id);
+            $product->delete();
+            return response()->json(['محصول با موفقیت حذف شد ']);
+        } else {
+            return response()->json(['message' => 'شما دسترسی لازم یرای انجام این کار را ندارید'], 403);
+        }
     }
 
-    public function restore($id)
+    public function restore(Request $request, $id)
     {
-        $product = Product::withTrashed()->findOrFail($id);
-        $product->restore();
-        return response()->json(['message' => 'محصول با موفقیت بازگردانی شد', 'product' => $product]);
+        if ($request->user()->can('product.restore')) {
+            $product = Product::withTrashed()->findOrFail($id);
+            $product->restore();
+            return response()->json(['message' => 'محصول با موفقیت بازگردانی شد', 'product' => $product]);
+        } else {
+            return response()->json(['message' => 'شما دسترسی لازم برای انجام این کار را ندارید'], 403);
+        }
     }
 }
